pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "deep-research-agent"
version = "0.1.0"
description = "Modular deep research automation pipeline"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Team" }]
dependencies = [
    "aiolimiter>=1.1",
    "anyio>=4.4",
    "fastapi>=0.110",
    "httpx[aio]>=0.27",
    "markdown>=3.5",
    "openai>=1.30",
    "pydantic>=2.6",
    "python-dotenv>=1.0",
    "sendgrid>=6.11",
    "tenacity>=8.2",
    "uvicorn[standard]>=0.30",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27",
    "pytest-asyncio>=0.23",
    "pytest>=8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

